+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 * Changes to some site configuration options are now automatically detected and no longer require a server restart. After hitting Save in the UI, you will be informed if a server restart is required, per usual.
 * Saved search notifications are now only sent to the owner of a saved search (all of an organization's members for an organization-level saved search, or a single user for a user-level saved search). The `notifyUsers` and `notifyOrganizations` properties underneath `search.savedQueries` have been removed.
 * Slack webhook URLs are now defined in user/organization JSON settings, not on the organization profile page. Previously defined organization Slack webhook URLs are automatically migrated to the organization's JSON settings.
+* The "unlimited" value for `maxReposToSearch` is now `-1` instead of `0`, and `0` now means to use the default.
 
 ### Bug fixes